extract xml extensions to XmlExtensions file
