+++ core/src/main/kotlin/com/malinskiy/marathon/report/junit/JUnitReporter.kt
@@ -64,25 +64,3 @@ class JUnitReporter(private val fileManager: FileManager) {
         writer.close()
     }
 }
-
-fun XMLStreamWriter.document(init: XMLStreamWriter.() -> Unit): XMLStreamWriter {
-    this.writeStartDocument()
-    this.init()
-    this.writeEndDocument()
-    return this
-}
-
-fun XMLStreamWriter.element(name: String, init: XMLStreamWriter.() -> Unit): XMLStreamWriter {
-    this.writeStartElement(name)
-    this.init()
-    this.writeEndElement()
-    return this
-}
-
-fun XMLStreamWriter.element(name: String, content: String) {
-    element(name) {
-        writeCharacters(content)
-    }
-}
-
-fun XMLStreamWriter.attribute(name: String, value: String) = writeAttribute(name, value)
